# apps/apis/src/modules/ingredients/dtos/create-ingredient.dto.ts
import { IsNotEmpty, IsNumber, IsString } from 'class-validator'

export class CreateIngredientDto {
    @IsString()
    @IsNotEmpty()
    name: string

    @IsString()
    @IsNotEmpty()
    image: string

    @IsNumber()
    @IsNotEmpty()
    carbohydrates: number

    @IsNumber()
    @IsNotEmpty()
    fat: number

    @IsNumber()
    @IsNotEmpty()
    protein: number
}