# apps/apis/src/modules/nutritional-values/dtos/update-nutritional-value.dto.ts
import { PartialType } from '@nestjs/swagger'
import { CreateNutritionalValueDto } from './create-nutritional-value.dto'

export class UpdateNutritionalValueDto extends PartialType(CreateNutritionalValueDto) {}