# apps/apis/swagger.json
{"openapi":"3.0.0","paths":{"/apis/todos/get-todos":{"get":{"operationId":"TodosController_getTodos","parameters":[],"responses":{"200":{"description":""}}}},"/apis/todos/get-todo/{id}":{"get":{"operationId":"TodosController_getTodo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/apis/todos/create-todo":{"post":{"operationId":"TodosController_createTodo","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":""}}}},"/apis/todos/update-todo/{id}":{"put":{"operationId":"TodosController_updateTodo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoDto"}}}},"responses":{"200":{"description":""}}}},"/apis/todos/delete-todo/{id}":{"delete":{"operationId":"TodosController_deleteTodo","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/apis/users/login":{"post":{"operationId":"UsersController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"]}},"/apis/users/refresh-token":{"post":{"operationId":"UsersController_refreshToken","parameters":[],"responses":{"201":{"description":""}},"tags":["users"]}},"/apis/users/login-with-google":{"get":{"operationId":"UsersController_loginWithGoogle","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/apis/users/google/callback":{"get":{"operationId":"UsersController_googleCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/apis/users/me":{"get":{"operationId":"UsersController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]}},"/apis/users/create-user":{"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]}},"/apis/users/user-info/{id}":{"get":{"operationId":"UsersController_getUserInfo","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the user","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/apis/meals":{"post":{"operationId":"MealsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMealDto"}}}},"responses":{"201":{"description":""}},"tags":["meals"]},"get":{"operationId":"MealsController_findAll","parameters":[{"name":"currentPage","required":false,"in":"query","description":"Trang hiện tại","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"Kích thước trang (số lượng bản ghi trên một trang)","schema":{"type":"number"}},{"name":"totalPage","required":false,"in":"query","description":"Tổng số trang (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}},{"name":"totalRecord","required":false,"in":"query","description":"Tổng số bản ghi (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["meals"]}},"/apis/meals/{id}":{"get":{"operationId":"MealsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["meals"]},"patch":{"operationId":"MealsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMealDto"}}}},"responses":{"200":{"description":""}},"tags":["meals"]},"delete":{"operationId":"MealsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["meals"]}},"/apis/meals/{id}/{userId}":{"patch":{"operationId":"MealsController_updateFavoriteMeal","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["meals"]}},"/apis/ingredients":{"post":{"operationId":"IngredientsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateIngredientDto"}}}},"responses":{"201":{"description":""}},"tags":["ingredients"]},"get":{"operationId":"IngredientsController_findAll","parameters":[{"name":"currentPage","required":false,"in":"query","description":"Trang hiện tại","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"Kích thước trang (số lượng bản ghi trên một trang)","schema":{"type":"number"}},{"name":"totalPage","required":false,"in":"query","description":"Tổng số trang (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}},{"name":"totalRecord","required":false,"in":"query","description":"Tổng số bản ghi (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResult"}}}}},"tags":["ingredients"]},"delete":{"operationId":"IngredientsController_deleteIngredients","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveIngredientsDto"}}}},"responses":{"200":{"description":""}},"tags":["ingredients"]}},"/apis/ingredients/recipe/{recipeId}":{"get":{"operationId":"IngredientsController_findAllByRecipeId","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}},{"name":"currentPage","required":false,"in":"query","description":"Trang hiện tại","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"Kích thước trang (số lượng bản ghi trên một trang)","schema":{"type":"number"}},{"name":"totalPage","required":false,"in":"query","description":"Tổng số trang (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}},{"name":"totalRecord","required":false,"in":"query","description":"Tổng số bản ghi (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["ingredients"]},"delete":{"operationId":"IngredientsController_deleteIngredientsByRecipe","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveIngredientsByRecipeDto"}}}},"responses":{"200":{"description":""}},"tags":["ingredients"]}},"/apis/ingredients/{id}":{"get":{"operationId":"IngredientsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ingredients"]},"patch":{"operationId":"IngredientsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateIngredientDto"}}}},"responses":{"200":{"description":""}},"tags":["ingredients"]},"delete":{"operationId":"IngredientsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ingredients"]}},"/apis/recipes":{"post":{"operationId":"RecipesController_create","parameters":[],"requestBody":{"required":true,"description":"Data of recipe to create new record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecipeDto"}}}},"responses":{"201":{"description":""}},"tags":["recipes"]},"get":{"operationId":"RecipesController_findAll","parameters":[{"name":"currentPage","required":false,"in":"query","description":"Trang hiện tại","schema":{"type":"number"}},{"name":"pageSize","required":false,"in":"query","description":"Kích thước trang (số lượng bản ghi trên một trang)","schema":{"type":"number"}},{"name":"totalPage","required":false,"in":"query","description":"Tổng số trang (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}},{"name":"totalRecord","required":false,"in":"query","description":"Tổng số bản ghi (không bắt buộc, thường được tính toán tự động)","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["recipes"]}},"/apis/recipes/{id}":{"get":{"operationId":"RecipesController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the recipe to find record","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["recipes"]},"patch":{"operationId":"RecipesController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the recipe to delete","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Data of recipe data to update","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRecipeDto"}}}},"responses":{"200":{"description":""}},"tags":["recipes"]},"delete":{"operationId":"RecipesController_delete","parameters":[{"name":"id","required":true,"in":"path","description":"Id of the recipe to delete","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["recipes"]}},"/apis/cuisines":{"post":{"operationId":"CuisinesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCuisineDto"}}}},"responses":{"201":{"description":""}},"tags":["cuisines"]},"get":{"operationId":"CuisinesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["cuisines"]}},"/apis/cuisines/{id}":{"get":{"operationId":"CuisinesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["cuisines"]},"patch":{"operationId":"CuisinesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCuisineDto"}}}},"responses":{"200":{"description":""}},"tags":["cuisines"]},"delete":{"operationId":"CuisinesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["cuisines"]}},"/apis/nutritional-values":{"post":{"operationId":"NutritionalValuesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNutritionalValueDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"NutritionalValuesController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/apis/nutritional-values/{id}":{"get":{"operationId":"NutritionalValuesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"NutritionalValuesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNutritionalValueDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"NutritionalValuesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"API","description":"### REST\n\nRoutes is following REST standard (Richardson level 3)\n\n<details><summary>Detailed specification</summary>\n<p>\n\n**List:**\n  - `GET /<resources>/`\n    - Get the list of **<resources>** as admin\n  - `GET /user/<user_id>/<resources>/`\n    - Get the list of **<resources>** for a given **<user_id>**\n    - Output a **403** if logged user is not **<user_id>**\n\n**Detail:**\n  - `GET /<resources>/<resource_id>`\n    - Get the detail for **<resources>** of id **<resource_id>**\n    - Output a **404** if not found\n  - `GET /user/<user_id>/<resources>/<resource_id>`\n    - Get the list of **<resources>** for a given **user_id**\n    - Output a **404** if not found\n    - Output a **403** if:\n      - Logged user is not **<user_id>**\n      - The **<user_id>** have no access to **<resource_id>**\n\n**Creation / Edition / Replacement / Suppression:**\n  - `<METHOD>` is:\n    - **POST** for creation\n    - **PATCH** for update (one or more fields)\n    - **PUT** for replacement (all fields, not used)\n    - **DELETE** for suppression (all fields, not used)\n  - `<METHOD> /<resources>/<resource_id>`\n    - Create **<resources>** with id **<resource_id>** as admin\n    - Output a **400** if **<resource_id>** conflicts with existing **<resources>**\n  - `<METHOD> /user/<user_id>/<resources>/<resource_id>`\n    - Create **<resources>** with id **<resource_id>** as a given **user_id**\n    - Output a **409** if **<resource_id>** conflicts with existing **<resources>**\n    - Output a **403** if:\n      - Logged user is not **<user_id>**\n      - The **<user_id>** have no access to **<resource_id>**\n</p>\n</details>","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateTodoDto":{"type":"object","properties":{}},"UpdateTodoDto":{"type":"object","properties":{}},"LoginUserDto":{"type":"object","properties":{"username":{"type":"string","example":"example@gmail.com"},"password":{"type":"string","example":"example"}},"required":["username","password"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","example":"example"},"email":{"type":"string","example":"example@gmail.com"},"password":{"type":"string","example":"example"}},"required":["username","email","password"]},"CreateMealDto":{"type":"object","properties":{}},"UpdateMealDto":{"type":"object","properties":{}},"CreateIngredientDto":{"type":"object","properties":{}},"PaginatedMetaResult":{"type":"object","properties":{"total":{"type":"number","description":"Total number of items"},"lastPage":{"type":"number","description":"Last page number"},"currentPage":{"type":"number","description":"Current page number"},"pageSize":{"type":"number","description":"Number of items per page"},"prevPage":{"type":"number","description":"Previous page number or null"},"nextPage":{"type":"number","description":"Next page number or null"}},"required":["total","lastPage","currentPage","pageSize","prevPage","nextPage"]},"PaginatedResult":{"type":"object","properties":{"data":{"description":"The data array","type":"array","items":{"type":"object"}},"meta":{"description":"Pagination metadata","allOf":[{"$ref":"#/components/schemas/PaginatedMetaResult"}]}},"required":["data","meta"]},"UpdateIngredientDto":{"type":"object","properties":{}},"RemoveIngredientsDto":{"type":"object","properties":{"ingredientIds":{"description":"List ingredient id that can be removes","type":"array","items":{"type":"string"}}},"required":["ingredientIds"]},"RemoveIngredientsByRecipeDto":{"type":"object","properties":{"ingredientIds":{"description":"List ingredient id that can be removes","type":"array","items":{"type":"string"}},"recipeId":{"type":"string","description":"Recipe is required"}},"required":["recipeId"]},"CreateRecipeDto":{"type":"object","properties":{"description":{"type":"string","description":"Description of the recipe","example":"Recipe example description."}},"required":["description"]},"UpdateRecipeDto":{"type":"object","properties":{"description":{"type":"string","description":"Description of the recipe","example":"Recipe example description."}}},"CreateCuisineDto":{"type":"object","properties":{}},"UpdateCuisineDto":{"type":"object","properties":{}},"CreateNutritionalValueDto":{"type":"object","properties":{}},"UpdateNutritionalValueDto":{"type":"object","properties":{}}}}}